requirements.txt
gunicorn
django==4.1.*
djangorestframework==3.14.*
django-filter
mysqlclient
coreapi